feat: Update Models for RAG Compatibility with MongoDB
Description:
We need to migrate our current application database from Supabase (PostgreSQL) to MongoDB. This involves transferring existing data, updating the backend to use MongoDB, and ensuring that all current features work seamlessly after migration.

Changelog:
Introduced beanie init that shall act as a centralized location that we can use to centralize the initialization of the Database and documents.

Since in Tortoise ORM postgresql we were using id of type UUID, but in MongoDB it usually uses _id with a binary version of UUID, in order to fix the transition from Postgresql(Tortoise ORM) to MongoDB (Beanie ODM) we need to introduce a Codec of type STANDARDIZED_UUID_CODEC, the reason we need it because MongoDB drivers historically wrote UUIDs using different binary encodings (“representations”: pythonLegacy, javaLegacy, csharpLegacy, etc.). If one service writes with one representation and another reads/queries with a different one, the same logical UUID becomes different bytes. which Effects:
        - find() by _id fails for existing documents
        - unique indexes allow duplicates
        - cross-language lookups and references break

This allows us to then declare id as the _id of type UUID in Beanie documents that will handle all the translations to UUID Binary and vice versa.