feat: Update Models for RAG Compatibility with MongoDB
Description:
We need to migrate our current application database from Supabase (PostgreSQL) to MongoDB. This involves transferring existing data, updating the backend to use MongoDB, and ensuring that all current features work seamlessly after migration.

Changelog:

In the old code every time I added a new method I had to go to several locations and add the mirror of this operation in Fakes, stubs, and even the validaiton was inconsistent, and all of this combined also hit the coverage hard. To solve these limitations, I managed to drop the complexity levels and reduced the boilerplate required everytime a new method is added making everything consistent, which made the coverage rates for tests powerful, with less boilerplate even more consistent, there is no more need to to repeat tests, I even automated certain operations that we had to do manually. The process of adding a new method is now Smooth

Usually what I have is:
- Specification, which is an interface that specifies the structure of the code and keeps everything consistent
- concrete ORM, ODM, and In memory backend implementation of the data storage backend that implements the Specification like the Tortoise, Beanie and Even In Memory for testing purposes Backends
- Base store that accepts backend implementations and takes responsibility for data validation